requirements.txt
cython-bbox==0.1.3
loguru==0.6.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
motmetrics==1.4.0
ninja==1.11.1

opencv-python-headless==4.9.0.80
pandas==1.5.3
Pillow==9.4.0
PyQt5==5.15.10

PyYAML==6.0

scikit-learn==1.2.1
scipy==1.9.1
seaborn==0.12.2

thop
tensorboard
lap
tabulate 
tqdm==4.65.0

ultralytics==8.0.94

wandb==0.15.0
